# PERF: Groupby aggregations with Categorical